refactor(tags): Separated converters and utility classes
